STAC-23268: move settings model generation to the settings extension
besides being a logical place to put it, now that the extension is there, it also prevents an awkward cyclic relationship between the tracetotopococonnector and the settingsproviderextension

# Conflicts:
#	extension/settingsproviderextension/internal/core/cache.go
#	extension/settingsproviderextension/internal/core/settings_helper.go
#	extension/settingsproviderextension/internal/core/subscriber.go
#	extension/settingsproviderextension/internal/core/subscriber_test.go
