+++ extension/settingsproviderextension/internal/core/cache.go
@@ -3,8 +3,8 @@ package core
 import (
 	"fmt"
 	"github.com/mohae/deepcopy"
-	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/connector/tracetotopoconnector/generated/settings"
 	stsSettingsEvents "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/events"
+	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/generated/settings"
 	"go.uber.org/zap"
 	"reflect"
 	"sync"
