+++ extension/settingsproviderextension/internal/core/settings_helper.go
@@ -2,7 +2,7 @@ package core
 
 import (
 	"fmt"
-	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/connector/tracetotopoconnector/generated/settings"
+	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/generated/settings"
 )
 
 type ConverterFunc func(stsSettingsModel.Setting) (any, error)
