+++ extension/settingsproviderextension/internal/core/settings_helper_test.go
@@ -1,7 +1,7 @@
 package core
 
 import (
-	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/connector/tracetotopoconnector/generated/settings"
+	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/generated/settings"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
