+++ extension/settingsproviderextension/internal/provider/file/file_settings_provider_test.go
@@ -2,8 +2,8 @@ package file
 
 import (
 	"context"
-	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/connector/tracetotopoconnector/generated/settings"
 	stsSettingsConfig "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/config"
+	stsSettingsModel "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/generated/settings"
 	stsSettingsCommon "github.com/stackvista/sts-opentelemetry-collector/extension/settingsproviderextension/internal/core"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
