+++ extension/settingsproviderextension/spec/settings.yaml
